fix(sprint4): DB 타입 불일치·하드코딩·마이그레이션 엣지케이스 수정
[DB 스키마]
- 09_migration_sprint4_fixes.sql 추가
  · playtime_analyses/critic_summaries: FLOAT → NUMERIC(5,2), game_id NOT NULL, id/game_id INTEGER → BIGINT
- 10_migration_sprint6_fixes.sql 수정
  · game_events CREATE TABLE 추가 (기존 ALTER TABLE만 있어 실행 시 오류 발생)
  · SERIAL/INTEGER → BIGSERIAL/BIGINT (games.id bigserial 참조 일관성)

[ORM 모델]
- PlaytimeAnalysis, CriticSummary: id/game_id Column(Integer) → Column(BigInteger)

[API]
- GET /api/v1/games/: tags(Steam platform_meta_json), rating(Metacritic 100→5점 환산) 필드 추가
- reviews.py: language_code 기본값 'ko' → 'en' (Metacritic/Steam 리뷰 언어 기본값 수정)

[프론트엔드]
- GameListPage, GameDetailPage: GAME_META 하드코딩 제거, API 응답(game.tags, game.rating) 사용
- rating/tags null-safe 처리 (API에 데이터 없으면 UI 요소 숨김)

+++ backend/app/models/domain.py
@@ -241,8 +241,8 @@ class PlaytimeAnalysis(Base):
     마이그레이션: 08_migration_sprint4.sql
     """
     __tablename__ = "playtime_analyses"
-    id                  = Column(Integer, primary_key=True, index=True)
-    game_id             = Column(Integer, ForeignKey("games.id"), nullable=False)
+    id                  = Column(BigInteger, primary_key=True, index=True)
+    game_id             = Column(BigInteger, ForeignKey("games.id"), nullable=False)
     bucket_thresholds   = Column(JSONB, nullable=False)
 
     early_summary       = Column(Text)
@@ -284,8 +284,8 @@ class CriticSummary(Base):
     마이그레이션: 08_migration_sprint4.sql
     """
     __tablename__ = "critic_summaries"
-    id              = Column(Integer, primary_key=True, index=True)
-    game_id         = Column(Integer, ForeignKey("games.id"), nullable=False)
+    id              = Column(BigInteger, primary_key=True, index=True)
+    game_id         = Column(BigInteger, ForeignKey("games.id"), nullable=False)
     summary         = Column(Text)
     sentiment       = Column(String(16))
     score           = Column(Numeric(5, 2))
